agents/requirements.txt
# MAGI System Dependencies

# Core frameworks
strands-agents>=1.0.0
bedrock-agentcore>=0.1.0
bedrock-agentcore-starter-toolkit>=0.1.0

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# Data validation
pydantic>=2.0.0

# Async support
aiohttp>=3.9.0
asyncio>=3.4.3

# Utilities
python-dateutil>=2.8.0
PyYAML>=6.0.0
